successful read of multilevel json
using the text object to compile lists of json entries; we can use double indices -- hopefully more -- for multi-dimensional arrays. perhaps this can be easier, but either way, this already seems to work better than the coll available though cyclone.

need to fix issue with third level (not adding to index)

also- will not load via [until] for some reason, so need to load individual lines on a (short) timer.