change im.helpfile to im.help
the example in im.acid-help.pd feels a bit more logical now...this looks more like it could be a central help resource that runs throughout the library rather than something I feel that have to hide away.

only issue is that the user has to lock the patcher to access these message boxes.