bring im.toggle closer new im.snapshot
would you consider this approach?
this way toggle looks like im.snapshot, either fully on or off, and we can still edit the name/arguments if desired.

while I was there, I added option of a bang to trigger the toggle if user is using 0.51+ (this prints a benign error in older versions)